Migrated nav list to a fragment
