+++ app/src/main/java/de/danoeh/antennapod/activity/MainActivity.java
@@ -1,24 +1,17 @@
 package de.danoeh.antennapod.activity;
 
 import android.annotation.TargetApi;
-import android.app.ActionBar;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
-import android.database.DataSetObserver;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
-import android.view.ContextMenu;
 import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
 import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -32,83 +25,48 @@ import androidx.fragment.app.FragmentTransaction;
 import com.bumptech.glide.Glide;
 import com.google.android.material.snackbar.Snackbar;
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.adapter.NavListAdapter;
-import de.danoeh.antennapod.core.asynctask.FeedRemover;
-import de.danoeh.antennapod.core.dialog.ConfirmationDialog;
-import de.danoeh.antennapod.core.event.FeedListUpdateEvent;
 import de.danoeh.antennapod.core.event.MessageEvent;
-import de.danoeh.antennapod.core.event.QueueEvent;
-import de.danoeh.antennapod.core.event.UnreadItemsUpdateEvent;
-import de.danoeh.antennapod.core.feed.Feed;
-import de.danoeh.antennapod.core.preferences.PlaybackPreferences;
 import de.danoeh.antennapod.core.preferences.UserPreferences;
-import de.danoeh.antennapod.core.service.playback.PlaybackService;
-import de.danoeh.antennapod.core.storage.DBReader;
-import de.danoeh.antennapod.core.storage.DBWriter;
-import de.danoeh.antennapod.core.util.FeedItemUtil;
 import de.danoeh.antennapod.core.util.Flavors;
-import de.danoeh.antennapod.core.util.IntentUtils;
 import de.danoeh.antennapod.core.util.StorageUtils;
 import de.danoeh.antennapod.dialog.RatingDialog;
-import de.danoeh.antennapod.dialog.RenameFeedDialog;
 import de.danoeh.antennapod.fragment.AddFeedFragment;
 import de.danoeh.antennapod.fragment.DownloadsFragment;
 import de.danoeh.antennapod.fragment.EpisodesFragment;
 import de.danoeh.antennapod.fragment.ExternalPlayerFragment;
 import de.danoeh.antennapod.fragment.FeedItemlistFragment;
+import de.danoeh.antennapod.fragment.NavDrawerFragment;
 import de.danoeh.antennapod.fragment.PlaybackHistoryFragment;
 import de.danoeh.antennapod.fragment.QueueFragment;
 import de.danoeh.antennapod.fragment.SubscriptionFragment;
 import de.danoeh.antennapod.fragment.TransitionEffect;
-import de.danoeh.antennapod.menuhandler.NavDrawerActivity;
 import de.danoeh.antennapod.preferences.PreferenceUpgrader;
-import io.reactivex.Observable;
-import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.disposables.Disposable;
-import io.reactivex.schedulers.Schedulers;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.Validate;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 
-import java.util.List;
-
 /**
  * The activity that is shown when the user launches the app.
  */
-public class MainActivity extends CastEnabledActivity implements NavDrawerActivity {
+public class MainActivity extends CastEnabledActivity {
 
     private static final String TAG = "MainActivity";
+    public static final String MAIN_FRAGMENT_TAG = "main";
 
     public static final String PREF_NAME = "MainActivityPrefs";
     public static final String PREF_IS_FIRST_LAUNCH = "prefMainActivityIsFirstLaunch";
-    public static final String PREF_LAST_FRAGMENT_TAG = "prefMainActivityLastFragmentTag";
 
-    public static final String EXTRA_NAV_TYPE = "nav_type";
-    public static final String EXTRA_NAV_INDEX = "nav_index";
     public static final String EXTRA_FRAGMENT_TAG = "fragment_tag";
     public static final String EXTRA_FRAGMENT_ARGS = "fragment_args";
     private static final String EXTRA_FEED_ID = "fragment_feed_id";
 
     private static final String SAVE_BACKSTACK_COUNT = "backstackCount";
 
-    public static final String[] NAV_DRAWER_TAGS = {
-            QueueFragment.TAG,
-            EpisodesFragment.TAG,
-            SubscriptionFragment.TAG,
-            DownloadsFragment.TAG,
-            PlaybackHistoryFragment.TAG,
-            AddFeedFragment.TAG,
-            NavListAdapter.SUBSCRIPTION_LIST_TAG
-    };
-
     private DrawerLayout drawerLayout;
     private View navDrawer;
-    private NavListAdapter navAdapter;
-    private int mPosition = -1;
     private ActionBarDrawerToggle drawerToggle;
-    private Disposable disposable;
     private long lastBackButtonPressTime = 0;
 
     @NonNull
@@ -127,39 +85,20 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         setContentView(R.layout.main);
 
         drawerLayout = findViewById(R.id.drawer_layout);
-        ListView navList = findViewById(R.id.nav_list);
-        navDrawer = findViewById(R.id.nav_layout);
+        navDrawer = findViewById(R.id.navDrawerFragment);
 
         final FragmentManager fm = getSupportFragmentManager();
         fm.addOnBackStackChangedListener(() ->
                 drawerToggle.setDrawerIndicatorEnabled(fm.getBackStackEntryCount() == 0));
 
-        navAdapter = new NavListAdapter(itemAccess, this);
-        navList.setAdapter(navAdapter);
-        navList.setOnItemClickListener(navListClickListener);
-        navList.setOnItemLongClickListener(newListLongClickListener);
-        registerForContextMenu(navList);
-
-        navAdapter.registerDataSetObserver(new DataSetObserver() {
-            @Override
-            public void onChanged() {
-                selectedNavListIndex = getSelectedNavListIndex();
-            }
-        });
-
-        findViewById(R.id.nav_settings).setOnClickListener(v -> {
-            drawerLayout.closeDrawer(navDrawer);
-            startActivity(new Intent(MainActivity.this, PreferenceActivity.class));
-        });
-
         FragmentTransaction transaction = fm.beginTransaction();
 
-        Fragment mainFragment = fm.findFragmentByTag("main");
+        Fragment mainFragment = fm.findFragmentByTag(MAIN_FRAGMENT_TAG);
         if (mainFragment != null) {
             transaction.replace(R.id.main_view, mainFragment);
         } else {
-            String lastFragment = getLastNavFragment();
-            if (ArrayUtils.contains(NAV_DRAWER_TAGS, lastFragment)) {
+            String lastFragment = NavDrawerFragment.getLastNavFragment(this);
+            if (ArrayUtils.contains(NavDrawerFragment.NAV_DRAWER_TAGS, lastFragment)) {
                 loadFragment(lastFragment, null);
             } else {
                 try {
@@ -174,6 +113,9 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         }
         ExternalPlayerFragment externalPlayerFragment = new ExternalPlayerFragment();
         transaction.replace(R.id.playerFragment, externalPlayerFragment, ExternalPlayerFragment.TAG);
+        NavDrawerFragment navDrawerFragment = new NavDrawerFragment();
+        transaction.replace(R.id.navDrawerFragment, navDrawerFragment, NavDrawerFragment.TAG);
+
         transaction.commit();
 
         checkFirstLaunch();
@@ -188,25 +130,6 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         drawerLayout.addDrawerListener(drawerToggle);
         drawerToggle.syncState();
         super.setSupportActionBar(toolbar);
-    }
-
-    private void saveLastNavFragment(String tag) {
-        Log.d(TAG, "saveLastNavFragment(tag: " + tag + ")");
-        SharedPreferences prefs = getSharedPreferences(PREF_NAME, MODE_PRIVATE);
-        SharedPreferences.Editor edit = prefs.edit();
-        if (tag != null) {
-            edit.putString(PREF_LAST_FRAGMENT_TAG, tag);
-        } else {
-            edit.remove(PREF_LAST_FRAGMENT_TAG);
-        }
-        edit.apply();
-    }
-
-    private String getLastNavFragment() {
-        SharedPreferences prefs = getSharedPreferences(PREF_NAME, MODE_PRIVATE);
-        String lastFragment = prefs.getString(PREF_LAST_FRAGMENT_TAG, QueueFragment.TAG);
-        Log.d(TAG, "getLastNavFragment() -> " + lastFragment);
-        return lastFragment;
     }
 
     private void checkFirstLaunch() {
@@ -224,50 +147,13 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         }
     }
 
-    private void showDrawerPreferencesDialog() {
-        final List<String> hiddenDrawerItems = UserPreferences.getHiddenDrawerItems();
-        String[] navLabels = new String[NAV_DRAWER_TAGS.length];
-        final boolean[] checked = new boolean[NAV_DRAWER_TAGS.length];
-        for (int i = 0; i < NAV_DRAWER_TAGS.length; i++) {
-            String tag = NAV_DRAWER_TAGS[i];
-            navLabels[i] = navAdapter.getLabel(tag);
-            if (!hiddenDrawerItems.contains(tag)) {
-                checked[i] = true;
-            }
-        }
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
-        builder.setTitle(R.string.drawer_preferences);
-        builder.setMultiChoiceItems(navLabels, checked, (dialog, which, isChecked) -> {
-            if (isChecked) {
-                hiddenDrawerItems.remove(NAV_DRAWER_TAGS[which]);
-            } else {
-                hiddenDrawerItems.add(NAV_DRAWER_TAGS[which]);
-            }
-        });
-        builder.setPositiveButton(R.string.confirm_label, (dialog, which) -> UserPreferences.setHiddenDrawerItems(hiddenDrawerItems));
-        builder.setNegativeButton(R.string.cancel_label, null);
-        builder.create().show();
-    }
-
     public boolean isDrawerOpen() {
         return drawerLayout != null && navDrawer != null && drawerLayout.isDrawerOpen(navDrawer);
     }
 
-    private void loadFragment(int index, Bundle args) {
-        Log.d(TAG, "loadFragment(index: " + index + ", args: " + args + ")");
-        if (index < navAdapter.getSubscriptionOffset()) {
-            String tag = navAdapter.getTags().get(index);
-            loadFragment(tag, args);
-        } else {
-            int pos = index - navAdapter.getSubscriptionOffset();
-            loadFeedFragmentByPosition(pos, args);
-        }
-    }
-
     public void loadFragment(String tag, Bundle args) {
         Log.d(TAG, "loadFragment(tag: " + tag + ", args: " + args + ")");
-        Fragment fragment = null;
+        Fragment fragment;
         switch (tag) {
             case QueueFragment.TAG:
                 fragment = new QueueFragment();
@@ -289,32 +175,22 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
                 break;
             default:
                 // default to the queue
-                tag = QueueFragment.TAG;
                 fragment = new QueueFragment();
                 args = null;
                 break;
         }
-        saveLastNavFragment(tag);
+
         if (args != null) {
             fragment.setArguments(args);
         }
         loadFragment(fragment);
     }
 
-    private void loadFeedFragmentByPosition(int relPos, Bundle args) {
-        if(relPos < 0) {
-            return;
-        }
-        Feed feed = itemAccess.getItem(relPos);
-        loadFeedFragmentById(feed.getId(), args);
-    }
-
     public void loadFeedFragmentById(long feedId, Bundle args) {
         Fragment fragment = FeedItemlistFragment.newInstance(feedId);
-        if(args != null) {
+        if (args != null) {
             fragment.setArguments(args);
         }
-        saveLastNavFragment(String.valueOf(feedId));
         loadFragment(fragment);
     }
 
@@ -325,7 +201,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
             fragmentManager.popBackStack();
         }
         FragmentTransaction t = fragmentManager.beginTransaction();
-        t.replace(R.id.main_view, fragment, "main");
+        t.replace(R.id.main_view, fragment, MAIN_FRAGMENT_TAG);
         fragmentManager.popBackStack();
         // TODO: we have to allow state loss here
         // since this function can get called from an AsyncTask which
@@ -334,9 +210,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         // not commit anything in an AsyncTask, but that's a bigger
         // change than we want now.
         t.commitAllowingStateLoss();
-        if (navAdapter != null) {
-            navAdapter.notifyDataSetChanged();
-        }
+        drawerLayout.closeDrawer(navDrawer);
     }
 
     public void loadChildFragment(Fragment fragment, TransitionEffect transition) {
@@ -357,8 +231,8 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         }
 
         transaction
-                .hide(getSupportFragmentManager().findFragmentByTag("main"))
-                .add(R.id.main_view, fragment, "main")
+                .hide(getSupportFragmentManager().findFragmentByTag(MAIN_FRAGMENT_TAG))
+                .add(R.id.main_view, fragment, MAIN_FRAGMENT_TAG)
                 .addToBackStack(null)
                 .commit();
     }
@@ -367,64 +241,10 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         loadChildFragment(fragment, TransitionEffect.NONE);
     }
 
-    private int getSelectedNavListIndex() {
-        String currentFragment = getLastNavFragment();
-        if(currentFragment == null) {
-            // should not happen, but better safe than sorry
-            return -1;
-        }
-        int tagIndex = navAdapter.getTags().indexOf(currentFragment);
-        if(tagIndex >= 0) {
-            return tagIndex;
-        } else if(ArrayUtils.contains(NAV_DRAWER_TAGS, currentFragment)) {
-            // the fragment was just hidden
-            return -1;
-        } else { // last fragment was not a list, but a feed
-            long feedId = Long.parseLong(currentFragment);
-            if (navDrawerData != null) {
-                List<Feed> feeds = navDrawerData.feeds;
-                for (int i = 0; i < feeds.size(); i++) {
-                    if (feeds.get(i).getId() == feedId) {
-                        return i + navAdapter.getSubscriptionOffset();
-                    }
-                }
-            }
-            return -1;
-        }
-    }
-
-    private final AdapterView.OnItemClickListener navListClickListener = new AdapterView.OnItemClickListener() {
-        @Override
-        public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-            int viewType = parent.getAdapter().getItemViewType(position);
-            if (viewType != NavListAdapter.VIEW_TYPE_SECTION_DIVIDER && position != selectedNavListIndex) {
-                loadFragment(position, null);
-            }
-            drawerLayout.closeDrawer(navDrawer);
-        }
-    };
-
-    private final AdapterView.OnItemLongClickListener newListLongClickListener = new AdapterView.OnItemLongClickListener() {
-        @Override
-        public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-            if(position < navAdapter.getTags().size()) {
-                showDrawerPreferencesDialog();
-                return true;
-            } else {
-                mPosition = position;
-                return false;
-            }
-        }
-    };
-
-
     @Override
     protected void onPostCreate(Bundle savedInstanceState) {
         super.onPostCreate(savedInstanceState);
         drawerToggle.syncState();
-        if (savedInstanceState != null) {
-            selectedNavListIndex = getSelectedNavListIndex();
-        }
     }
 
     @Override
@@ -455,14 +275,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
     protected void onResume() {
         super.onResume();
         StorageUtils.checkStorageAvailability(this);
-
-        Intent intent = getIntent();
-        if (intent.hasExtra(EXTRA_FEED_ID) ||
-                (navDrawerData != null && intent.hasExtra(EXTRA_NAV_TYPE) &&
-                        (intent.hasExtra(EXTRA_NAV_INDEX) || intent.hasExtra(EXTRA_FRAGMENT_TAG)))) {
-            handleNavIntent();
-        }
-        loadData();
+        handleNavIntent();
         RatingDialog.check();
     }
 
@@ -470,11 +283,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
     protected void onStop() {
         super.onStop();
         EventBus.getDefault().unregister(this);
-        if (disposable != null) {
-            disposable.dispose();
-        }
     }
-
 
     @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
     @Override
@@ -493,7 +302,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
     public boolean onCreateOptionsMenu(Menu menu) {
         boolean retVal = super.onCreateOptionsMenu(menu);
         if (Flavors.FLAVOR == Flavors.PLAY) {
-            switch (getLastNavFragment()) {
+            switch (NavDrawerFragment.getLastNavFragment(this)) {
                 case QueueFragment.TAG:
                 case EpisodesFragment.TAG:
                     requestCastButton(MenuItem.SHOW_AS_ACTION_IF_ROOM);
@@ -526,100 +335,6 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         }
     }
 
-
-    @Override
-    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
-        super.onCreateContextMenu(menu, v, menuInfo);
-        if(v.getId() != R.id.nav_list) {
-            return;
-        }
-        AdapterView.AdapterContextMenuInfo adapterInfo = (AdapterView.AdapterContextMenuInfo) menuInfo;
-        int position = adapterInfo.position;
-        if(position < navAdapter.getSubscriptionOffset()) {
-            return;
-        }
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.nav_feed_context, menu);
-        Feed feed = navDrawerData.feeds.get(position - navAdapter.getSubscriptionOffset());
-        menu.setHeaderTitle(feed.getTitle());
-        // episodes are not loaded, so we cannot check if the podcast has new or unplayed ones!
-    }
-
-
-    @Override
-    public boolean onContextItemSelected(MenuItem item) {
-        final int position = mPosition;
-        mPosition = -1; // reset
-        if(position < 0) {
-            return false;
-        }
-        Feed feed = navDrawerData.feeds.get(position - navAdapter.getSubscriptionOffset());
-        switch(item.getItemId()) {
-            case R.id.remove_all_new_flags_item:
-                ConfirmationDialog removeAllNewFlagsConfirmationDialog = new ConfirmationDialog(this,
-                        R.string.remove_all_new_flags_label,
-                        R.string.remove_all_new_flags_confirmation_msg) {
-                    @Override
-                    public void onConfirmButtonPressed(DialogInterface dialog) {
-                        dialog.dismiss();
-                        DBWriter.removeFeedNewFlag(feed.getId());
-                    }
-                };
-                removeAllNewFlagsConfirmationDialog.createNewDialog().show();
-                return true;
-            case R.id.mark_all_read_item:
-                ConfirmationDialog markAllReadConfirmationDialog = new ConfirmationDialog(this,
-                        R.string.mark_all_read_label,
-                        R.string.mark_all_read_confirmation_msg) {
-
-                    @Override
-                    public void onConfirmButtonPressed(DialogInterface dialog) {
-                        dialog.dismiss();
-                        DBWriter.markFeedRead(feed.getId());
-                    }
-                };
-                markAllReadConfirmationDialog.createNewDialog().show();
-                return true;
-            case R.id.rename_item:
-                new RenameFeedDialog(this, feed).show();
-                return true;
-            case R.id.remove_item:
-                final FeedRemover remover = new FeedRemover(this, feed) {
-                    @Override
-                    protected void onPostExecute(Void result) {
-                        super.onPostExecute(result);
-                        if(getSelectedNavListIndex() == position) {
-                            loadFragment(EpisodesFragment.TAG, null);
-                        }
-                    }
-                };
-                ConfirmationDialog conDialog = new ConfirmationDialog(this,
-                        R.string.remove_feed_label,
-                        getString(R.string.feed_delete_confirmation_msg, feed.getTitle())) {
-                    @Override
-                    public void onConfirmButtonPressed(
-                            DialogInterface dialog) {
-                        dialog.dismiss();
-                        long mediaId = PlaybackPreferences.getCurrentlyPlayingFeedMediaId();
-                        if (mediaId > 0 &&
-                                FeedItemUtil.indexOfItemWithMediaId(feed.getItems(), mediaId) >= 0) {
-                            Log.d(TAG, "Currently playing episode is about to be deleted, skipping");
-                            remover.skipOnCompletion = true;
-                            int playerStatus = PlaybackPreferences.getCurrentPlayerStatus();
-                            if(playerStatus == PlaybackPreferences.PLAYER_STATUS_PLAYING) {
-                                IntentUtils.sendLocalBroadcast(MainActivity.this, PlaybackService.ACTION_PAUSE_PLAY_CURRENT_EPISODE);
-                            }
-                        }
-                        remover.executeAsync();
-                    }
-                };
-                conDialog.createNewDialog().show();
-                return true;
-            default:
-                return super.onContextItemSelected(item);
-        }
-    }
-
     @Override
     public void onBackPressed() {
         if (isDrawerOpen()) {
@@ -648,7 +363,7 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
                     }
                     break;
                 case GO_TO_PAGE:
-                    if (getLastNavFragment().equals(UserPreferences.getBackButtonGoToPage())) {
+                    if (NavDrawerFragment.getLastNavFragment(this).equals(UserPreferences.getBackButtonGoToPage())) {
                         super.onBackPressed();
                     } else {
                         loadFragment(UserPreferences.getBackButtonGoToPage(), null);
@@ -659,141 +374,32 @@ public class MainActivity extends CastEnabledActivity implements NavDrawerActivi
         }
     }
 
-    private DBReader.NavDrawerData navDrawerData;
-    private int selectedNavListIndex = 0;
-
-    private final NavListAdapter.ItemAccess itemAccess = new NavListAdapter.ItemAccess() {
-        @Override
-        public int getCount() {
-            if (navDrawerData != null) {
-                return navDrawerData.feeds.size();
-            } else {
-                return 0;
-            }
-        }
-
-        @Override
-        public Feed getItem(int position) {
-            if (navDrawerData != null && 0 <= position && position < navDrawerData.feeds.size()) {
-                return navDrawerData.feeds.get(position);
-            } else {
-                return null;
-            }
-        }
-
-        @Override
-        public int getSelectedItemIndex() {
-            return selectedNavListIndex;
-        }
-
-        @Override
-        public int getQueueSize() {
-            return (navDrawerData != null) ? navDrawerData.queueSize : 0;
-        }
-
-        @Override
-        public int getNumberOfNewItems() {
-            return (navDrawerData != null) ? navDrawerData.numNewItems : 0;
-        }
-
-        @Override
-        public int getNumberOfDownloadedItems() {
-            return (navDrawerData != null) ? navDrawerData.numDownloadedItems : 0;
-        }
-
-        @Override
-        public int getReclaimableItems() {
-            return (navDrawerData != null) ? navDrawerData.reclaimableSpace : 0;
-        }
-
-        @Override
-        public int getFeedCounter(long feedId) {
-            return navDrawerData != null ? navDrawerData.feedCounters.get(feedId) : 0;
-        }
-
-        @Override
-        public int getFeedCounterSum() {
-            if(navDrawerData == null) {
-                return 0;
-            }
-            int sum = 0;
-            for(int counter : navDrawerData.feedCounters.values()) {
-                sum += counter;
-            }
-            return sum;
-        }
-
-    };
-
-    private void loadData() {
-        disposable = Observable.fromCallable(DBReader::getNavDrawerData)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(result -> {
-                    boolean handleIntent = (navDrawerData == null);
-
-                    navDrawerData = result;
-                    navAdapter.notifyDataSetChanged();
-
-                    if (handleIntent) {
-                        handleNavIntent();
-                    }
-                }, error -> Log.e(TAG, Log.getStackTraceString(error)));
-    }
-
-    @Subscribe
-    public void onEvent(QueueEvent event) {
-        Log.d(TAG, "onEvent(" + event + ")");
-        // we are only interested in the number of queue items, not download status or position
-        if(event.action == QueueEvent.Action.DELETED_MEDIA ||
-                event.action == QueueEvent.Action.SORTED ||
-                event.action == QueueEvent.Action.MOVED) {
-            return;
-        }
-        loadData();
-    }
-
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onEventMainThread(MessageEvent event) {
         Log.d(TAG, "onEvent(" + event + ")");
         View parentLayout = findViewById(R.id.drawer_layout);
         Snackbar snackbar = Snackbar.make(parentLayout, event.message, Snackbar.LENGTH_SHORT);
-        if(event.action != null) {
+        if (event.action != null) {
             snackbar.setAction(getString(R.string.undo), v -> event.action.run());
         }
         snackbar.show();
     }
 
-    @Subscribe
-    public void onUnreadItemsChanged(UnreadItemsUpdateEvent event) {
-        loadData();
-    }
-
-
-    @Subscribe
-    public void onFeedListChanged(FeedListUpdateEvent event) {
-        loadData();
-    }
-
     private void handleNavIntent() {
-        Log.d(TAG, "handleNavIntent()");
         Intent intent = getIntent();
-        if (intent.hasExtra(EXTRA_FEED_ID) ||
-                (intent.hasExtra(EXTRA_NAV_TYPE) &&
-                        (intent.hasExtra(EXTRA_NAV_INDEX) || intent.hasExtra(EXTRA_FRAGMENT_TAG)))) {
-            int index = intent.getIntExtra(EXTRA_NAV_INDEX, -1);
+        if (intent.hasExtra(EXTRA_FEED_ID) || intent.hasExtra(EXTRA_FRAGMENT_TAG)) {
+            Log.d(TAG, "handleNavIntent()");
             String tag = intent.getStringExtra(EXTRA_FRAGMENT_TAG);
             Bundle args = intent.getBundleExtra(EXTRA_FRAGMENT_ARGS);
             long feedId = intent.getLongExtra(EXTRA_FEED_ID, 0);
-            if (index >= 0) {
-                loadFragment(index, args);
-            } else if (tag != null) {
+            if (tag != null) {
                 loadFragment(tag, args);
-            } else if(feedId > 0) {
+            } else if (feedId > 0) {
                 loadFeedFragmentById(feedId, args);
             }
+            // to avoid handling the intent twice when the configuration changes
+            setIntent(new Intent(MainActivity.this, MainActivity.class));
         }
-        setIntent(new Intent(MainActivity.this, MainActivity.class)); // to avoid handling the intent twice when the configuration changes
     }
 
     @Override
