+++ app/src/main/java/de/danoeh/antennapod/config/DownloadServiceCallbacksImpl.java
@@ -19,14 +19,11 @@ public class DownloadServiceCallbacksImpl implements DownloadServiceCallbacks {
     @Override
     public PendingIntent getNotificationContentIntent(Context context) {
         Intent intent = new Intent(context, MainActivity.class);
-        intent.putExtra(MainActivity.EXTRA_NAV_TYPE, NavListAdapter.VIEW_TYPE_NAV);
         intent.putExtra(MainActivity.EXTRA_FRAGMENT_TAG, DownloadsFragment.TAG);
         Bundle args = new Bundle();
         args.putInt(DownloadsFragment.ARG_SELECTED_TAB, DownloadsFragment.POS_RUNNING);
         intent.putExtra(MainActivity.EXTRA_FRAGMENT_ARGS, args);
-
-        return PendingIntent.getActivity(context, 0, intent,
-                PendingIntent.FLAG_UPDATE_CURRENT);
+        return PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
     }
 
     @Override
@@ -40,7 +37,6 @@ public class DownloadServiceCallbacksImpl implements DownloadServiceCallbacks {
     @Override
     public PendingIntent getReportNotificationContentIntent(Context context) {
         Intent intent = new Intent(context, MainActivity.class);
-        intent.putExtra(MainActivity.EXTRA_NAV_TYPE, NavListAdapter.VIEW_TYPE_NAV);
         intent.putExtra(MainActivity.EXTRA_FRAGMENT_TAG, DownloadsFragment.TAG);
         Bundle args = new Bundle();
         args.putInt(DownloadsFragment.ARG_SELECTED_TAB, DownloadsFragment.POS_LOG);
