--- app/src/main/java/de/danoeh/antennapod/menuhandler/NavDrawerActivity.java
@@ -1,9 +1,0 @@
-package de.danoeh.antennapod.menuhandler;
-
-/**
- * Defines useful methods for activities that have a navigation drawer
- */
-public interface NavDrawerActivity {
-
-    boolean isDrawerOpen();
-}
