+++ app/src/main/res/layout/main.xml
@@ -30,6 +30,14 @@
 
     </RelativeLayout>
 
-    <include layout="@layout/nav_list" />
+    <FrameLayout
+        android:id="@+id/navDrawerFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginEnd="24dp"
+        android:layout_marginRight="24dp"
+        android:layout_gravity="start"
+        android:background="?android:attr/windowBackground"
+        android:orientation="vertical" />
 
 </androidx.drawerlayout.widget.DrawerLayout>
